refactor secq
