+++ algebra/src/secp256k1/fq.rs
@@ -1,0 +1,4 @@
+use crate::secq256k1::SECQ256K1Scalar;
+
+/// The wrapped struct for `ark_bulletproofs::curve::secp256k1::Fq`
+pub type SECP256K1Fq = SECQ256K1Scalar;
