+++ algebra/src/secq256k1/fq.rs
@@ -1,0 +1,4 @@
+use crate::secp256k1::SECP256K1Scalar;
+
+/// The wrapped struct for `ark_bulletproofs::curve::secq256k1::Fq`
+pub type SECQ256K1Fq = SECP256K1Scalar;
